DASH_Userguides_Assistant/requirements.txt
langchain
langchain-openai
langchain-community
langchain-huggingface
langchain-text-splitters
langchain-groq
langchain-chroma
litellm
pypdf
docx2txt
google-genai
google-adk
sentence-transformers
chromadb
pydantic
python-dotenv